Move `StoryComment` to its own file - Also rename `Story Comment` module to `Story Details` to give more context of the feature
