+++ HackrNewsiOS/Story UI/Controllers/StoryViewController.swift
@@ -2,27 +2,8 @@
 //  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
 //
 
+import HackrNews
 import UIKit
-
-public struct StoryDetail {
-    public let title: String
-    public let author: String
-    public let score: Int
-    public let createdAt: Date
-    public let totalComments: Int
-    public let comments: [Int]
-    public let url: URL
-
-    public init(title: String, author: String, score: Int, createdAt: Date, totalComments: Int, comments: [Int], url: URL) {
-        self.title = title
-        self.author = author
-        self.score = score
-        self.createdAt = createdAt
-        self.totalComments = totalComments
-        self.comments = comments
-        self.url = url
-    }
-}
 
 public class StoryViewController: UITableViewController {
     private var locale = Locale.current
